Restart Neovim configuration
I'm making a clean Neovim configuration due to better understanding
of the documentation and how the Lazy package manager works.

As of know this is a minimal configuration with only plugins for:

* colorscheme (onedark)
* nvim-treesitter
* lua-line

The plan is to expand the configuration as is needed and have a
plan for the "final" form.